requirements.txt
hopsworks[python,great-expectations]>=4.3.1
streamlit==1.28.2
email-validator==2.2.0
pydantic-settings>=2.6.1
requests-cache==1.2.0
retry-requests==2.0.0
xgboost==2.0.3
scikit-learn==1.2.2
matplotlib==3.8.3
plotly
seaborn
nbformat
Faker
invoke
python-dotenv
pandas>=2.0.0        
feldera==0.41
xgboost==2.0.3
scikit-learn==1.2.2

